src/main/webapp/app/entities/horario-consulta/delete/horario-consulta-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { HorarioConsultaService } from '../service/horario-consulta.service';

import { HorarioConsultaDeleteDialogComponent } from './horario-consulta-delete-dialog.component';

describe('HorarioConsulta Management Delete Component', () => {
  let comp: HorarioConsultaDeleteDialogComponent;
  let fixture: ComponentFixture<HorarioConsultaDeleteDialogComponent>;
  let service: HorarioConsultaService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HorarioConsultaDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(HorarioConsultaDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(HorarioConsultaDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(HorarioConsultaService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
